Restructured the Fortran code and added all the missing projects form Kirkup's web page. Also adding makefiles for the projects. Some projects do not yet work. The 2D external example craps out with a warning message about coincident vertices.
All the Modal analysis examples don't work yet, because they rely on a deprecated NAG routine
for eignvalue decomposition. This should not be a big deal as I think this will be handled in
the Python code I'm going to port this to. For testing purposes, I think I'll be able to replace
this with code from the GNU scientific library, as I previously did with the Hankel function code.